chore: refactor strategies; better addRooms options
